test: clear the two lint breaks current main hands to every PR
The core lint stripe fails on src/infra/state-migrations.media-persistence.test.ts
(1034 counted lines against the 1000 max), and the lint-suppressions allowlist
records one preserve-caught-error entry for extensions/qa-lab/src/gateway-child-setup.ts
while the tree has two. Both break every PR run that pulls either lane and
neither is owned by a pending change.

The media persistence fixtures (temp dirs, legacy database builder, snapshot
reader, archive writer) move to a colocated fixture module, matching the
existing run.overflow-compaction.fixture.ts precedent; the test cases and their
lane are unchanged.

The second gateway-child-setup suppression is the intentional redaction
boundary at the plugin setup failure path: candidate output can quote
credentials, so the error cause is dropped and only the bounded redacted
diagnostic crosses. The allowlist records it with that justification instead
of weakening the boundary.
